Add cast to uint8_t when checking for BOM

+++ include/xml.hpp
@@ -318,7 +318,7 @@ namespace xml
             static bool hasByteOrderMark(Iterator begin, Iterator end) noexcept
             {
                 for (auto i = std::begin(utf8ByteOrderMark); i != std::end(utf8ByteOrderMark); ++i, ++begin)
-                    if (begin == end || *begin != *i)
+                    if (begin == end || static_cast<std::uint8_t>(*begin) != *i)
                         return false;
                 return true;
             }
